refactor: extract OpenAI ingress tests
